Add functionality for downloading employee data #38
