+++ employee-skills-ui/src/App.js
@@ -1,15 +1,13 @@
 import './App.css';
 import Navbar from './components/nav-bar/Navbar'
 import AgGrid from './agGrid'
-import { Container, Grid, ThemeProvider, CssBaseline, SpeedDial, SpeedDialIcon, SpeedDialAction } from '@mui/material';
+import { Container, Grid, ThemeProvider, CssBaseline} from '@mui/material';
 import { useSelector } from 'react-redux';
 import {  LIGHT_THEME } from './redux/theme/themeConstants';
 import {
   createBrowserRouter,
   RouterProvider,
 } from "react-router-dom";
-import SaveIcon from '@mui/icons-material/Save';
-import FileUploadIcon from '@mui/icons-material/FileUpload';
 import EmployeeDetailContainer from './components/employee-detail/EmployeeDetailContainer';
 import React from 'react';
 import { getThemeObj } from './utils/utils';
@@ -28,10 +26,7 @@ const router = createBrowserRouter([
 
 
 function App(props) {
-  const actions = [
-    { icon: <FileUploadIcon />, name: 'Upload CSV' },
-    { icon: <SaveIcon />, name: 'Download' },
-  ];
+
 
 
   const mode = useSelector(state => state.theme.mode);
@@ -49,26 +44,8 @@ function App(props) {
 
         <Grid container spacing={2} >
           <Grid item xs={12} md={12}>
-          <RouterProvider router={router} />
+            <RouterProvider router={router} />
           </Grid>
-          <SpeedDial
-            ariaLabel="SpeedDial basic example"
-            sx={{ position: 'absolute', bottom: 16, right: 16,'& .MuiFab-primary': { backgroundColor: 'secondary.main' } }}
-            icon={<SpeedDialIcon />}
-            
-          >
-            {actions.map((action) => (
-              <SpeedDialAction
-                key={action.name}
-                icon={action.icon}
-                tooltipTitle={action.name}
-              />
-            ))}
-          </SpeedDial>
-          <input
-    type="file"
-    hidden
-  />
         </Grid>
       </Container>
       </Container>
